requirements.txt
pytorch-lightning==2.1.3
matplotlib
protobuf==3.19
scipy
Shapely==1.8.0
tensorflow
torchmetrics
tqdm==4.62.3
pyyaml
imageio
wandb
yapf==0.30.0
opencv-python
waymo-open-dataset-tf-2-11-0==1.6.0
ray
natten==0.14.6
timm==0.9.12
gradio==3.45.0
bidict